pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "telos-runtime"
version = "0.1.0-alpha"
description = "An open-source runtime where the AI decides what to build next."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Tomh" }
]
dependencies = [
    "litellm>=1.0.0",
    "qdrant-client>=1.7.0",
    "sqlalchemy>=2.0.0",
    "docker>=7.0.0",
    "click>=8.1.0",
    "pyyaml>=6.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0"
]

[project.scripts]
telos = "telos.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/telos"]
